feat(hpa): adding scaling policies for scaling down
The HPA mechanism allow to create specific policies to scale up/down the pods with specific timing so to avoid swift changes on the number of resources used.
This feature is adding two scale down policies to avoid resources to be reduced too quickly:
- only 1 pod every 60 seconds can be dismissed
- only the 15% of the pods every 60 seconds can be dismissed
The policy that has an higher impact on the pod number will trigger first.

Refs: Activiti/Activiti#3791