Create a template to support for Horizontal Pod Autoscaling (HPA) (#37)
* feat(hpa): create a template to support for Horizontal Autoscaling Pods

added a new _hpa.tpl template with unit tests

Refs: Activiti/Activiti#3791

* feat(hpa): adding scaling policies for scaling down

The HPA mechanism allow to create specific policies to scale up/down the pods with specific timing so to avoid swift changes on the number of resources used.
This feature is adding two scale down policies to avoid resources to be reduced too quickly:
- only 1 pod every 60 seconds can be dismissed
- only the 15% of the pods every 60 seconds can be dismissed
The policy that has an higher impact on the pod number will trigger first.

Refs: Activiti/Activiti#3791

* feat(hpa): added the `scalingPolicesEnabled` flag to opt out the polices

The following configuration:
```
runtime-bundle:
  hpa:
    scalingPolicesEnabled: false
```
will exclude the HPA scale down polices to be applied.
if `scalingPolicesEnabled` is true or absent they will.

Ref: Activiti/Activiti#3791

* fix(hpa): hpa must differentiate among Deployment and StatfulSet

Refs: Activiti/Activiti#3791

* fix(hpa): duplicated test case fixed

The test case was erroneously duplicated. Now there are two tests for:
- enabled: false
- enabled: true and enables.hpa: false

Refs: Activiti/Activiti#3791

* fix(hpa): improving a test name

Refs: Activiti/Activiti#3791

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>